catch all errors for extract doi
it appears many different errors could occur while parsing the url
for dois.  To avoid the ingestion failing on a weird url, we are just
logging the error and setting the doi to null.  That way, during the
validation phase the record will be considered invalid and the
ingestion can continue